Functions for interpolation/calculation of gravity etc. moved to new module gravity.py.
