+++ src/fire/api/geodetic_levelling/__init__.py
@@ -5,6 +5,7 @@ geodetic_levelling
 import fire.api.geodetic_levelling.contour_plot
 import fire.api.geodetic_levelling.geodetic_correction_levelling_obs
 import fire.api.geodetic_levelling.histogram
+import fire.api.geodetic_levelling.gravity
 import fire.api.geodetic_levelling.metric_to_gpu_transformation
 import fire.api.geodetic_levelling.tidal_transformation
 import fire.api.geodetic_levelling.time_propagation
